Fix type definition for custom PyObjects (static PyTypeObject)
The definition for the structure PyTypeObject has changed.
To properly function, custom type structures must be initialized
with PyType_Ready(), function that requires types to have
a valid name. Consequently, the zero between the head and the
name has to be removed.

An IntializeTypes() function was created to initialize all
types. The module won't be properly initialized if
any of its types fails to be ready.